+++ CHANGELOG.md
@@ -1,4 +1,19 @@
-# HTTP Change Log
+# Yii HTTP Change Log
+
+
+## 1.1.2 under development
+
+- no changes in this release.
+
+
+## 1.1.1 February 10, 2021
+
+- Chg: Update yiisoft/strings dependency (samdark)
+
+## 1.1.0 December 28, 2020
+
+- Enh #12: Add `Method::ALL` and deprecated `Method::ANY` (samdark)
+- Enh #20: Add `ContentDispositionHeader` that generate `Content-Disposition` header name and value (vjik)
 
 ## 1.0.0 September 1, 2020